+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bsv-bap",
-  "version": "0.1.4",
+  "version": "0.1.6",
   "description": "BAP npm module",
   "repository": {
     "type": "git",
